fix(window): FIRST_VALUE/LAST_VALUE/NTH_VALUE frame boundary handling (#5038)
* fix(window): make FIRST_VALUE/LAST_VALUE/NTH_VALUE respect frame boundaries

FIRST_VALUE, LAST_VALUE, and NTH_VALUE were computing a single value
from the entire partition and broadcasting it to every row. They never
called calculate_frame_with_exclusion, ignoring the window frame spec.

Refactored all three functions in both dispatch sites (evaluation.rs and
aggregation/window.rs) to loop per-row and compute frame boundaries,
matching the pattern used by aggregate window functions. This correctly
handles default frames (RANGE BETWEEN UNBOUNDED PRECEDING AND CURRENT
ROW), explicit frame specs, and EXCLUDE clauses.

window3.test: 993 passed -> 1138 passed (+145 tests)

Closes #5032


* fix(test): update test_window_last_value for correct frame-aware behavior

LAST_VALUE with default frame (ROWS BETWEEN UNBOUNDED PRECEDING AND
CURRENT ROW) should return the current row's value, not the
partition-wide last value.


---------

+++ tests/executor/window_functions.rs
@@ -564,14 +564,15 @@ fn test_window_last_value() {
         // Should have 4 rows
         assert_eq!(result.len(), 4);
 
-        // Engineering rows should have lowest_salary = 95000 (Bob is last in DESC order)
-        assert_eq!(result[0].values[3], SqlValue::Integer(95000));
-        assert_eq!(result[1].values[3], SqlValue::Integer(95000));
+        // With default frame (ROWS BETWEEN UNBOUNDED PRECEDING AND CURRENT ROW),
+        // LAST_VALUE returns the current row's value (last in the frame so far)
+        // Engineering: Alice (120000) then Bob (95000) in DESC order
+        assert_eq!(result[0].values[3], SqlValue::Integer(120000)); // frame: [Alice] → 120000
+        assert_eq!(result[1].values[3], SqlValue::Integer(95000));  // frame: [Alice, Bob] → 95000
 
-        // Sales rows should have lowest_salary = 85000 (Carol is last in DESC order)
-        // When sorted by salary DESC: Dave (90000) comes before Carol (85000)
-        assert_eq!(result[2].values[3], SqlValue::Integer(85000));
-        assert_eq!(result[3].values[3], SqlValue::Integer(85000));
+        // Sales: Dave (90000) then Carol (85000) in DESC order
+        assert_eq!(result[2].values[3], SqlValue::Integer(90000));  // frame: [Dave] → 90000
+        assert_eq!(result[3].values[3], SqlValue::Integer(85000));  // frame: [Dave, Carol] → 85000
 
         println!("✅ LAST_VALUE test works!");
     } else {
